EECS Game: First Commit
This commit contains the basic components
for EECS: The Game, a 3D combat game based
around mathematics, electrical and software
engineering. This commit contains a working
version of a Mathematician character and the
starting bones of an Electrical Engineer
character.

Players can run, jump, backup, and crouch.
The current weapons include Integrals,
Summations, and Vectors with more weapons
to be added later.
(See Player One controls in README)

The 2-player function has not been implemented.
The intention is to create an online multiplayer
so that players don't need to use the same keyboard
and/or gamepad.
The default controls are KeyboardMouse
unless the player opts for another implemented
control scheme (Keyboard/Gamepad).