Player Prefab: Fighting System Reduction
This commit contains the following updates, changes, and fixes:

The kick and punch logic for both player types has been reduced
to only support single attacks instead of combos and the
effectiveness of the attack has been delayed to line up with the stun
animation and health decrement of the opponent.
The logic is also able to accomodate button mashing by the player
without delays or additional opponent health decrements.

The special attack animation for both players has been checked
for "Has Exit Time" to keep the animation playing to completion
during the attack.

The Mathematician's Vector Attack duration has been extended. A bug
was discovered that led to the Vector Attack being shortened
significantly, but this update has fixed that issue.
A similar bug was found with the Electrical Engineer's clock attack
and it was resolved with the same fix.

The Integral can no longer cause damage to the opponent
when the player is not attacking, even if a previous attack missed.

The player controller script has been partitioned based on which
methods and variables are used by which players in preparation to
split the script into two separate scripts for the individual playable
characters at a later date.