EECS Game: UI + Bug Fixes
This commit generates the support for UI
interaction with the main menu and the
player select screens via the controller.

This is a base for the screens and they
will be updated over time.

Bug Fix: Health Bar from previous game was
not destroyed and did not register with the
players in the new game.

Bug Fix: UI re-appeared and took inputs during
an active game.